fixed error: call of overloaded 'abs(float)' is ambiguous
git-svn-id: svn://maxr.org/MAX/trunk@2394 c3bd5e88-ca0b-48cf-b972-da51c77c4c1b

+++ hud.cpp
@@ -2574,7 +2574,7 @@ void cGameGUI::miniMapMovedOver( void *parent )
 void cGameGUI::zoomSliderMoved( void *parent )
 {
 	cGameGUI *gui = static_cast<cGameGUI*>(parent);
-	gui->setZoom ( abs( gui->zoomSlider->getValue() ), false );
+	gui->setZoom ( abs( (int)gui->zoomSlider->getValue() ), false );
 }
 
 void cGameGUI::endReleased( void *parent )
